\name{Rcplex.close}
\alias{Rcplex.close}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Release CPLEX license }
\description{
  This function releases the currently held CPLEX license.
}
\usage{
Rcplex.close()
}
\author{Hector Corrada Bravo }
\seealso{\code{\link{Rcplex}} }
\keyword{utilities}
